Fix stale/duplicate latest heartbeat and retention overflow
- getLatestHeartbeatPerDevice: key on MAX(timestamp) instead of
  MAX(receivedAt). receivedAt is the receiver's insert clock, so during a
  catch-up burst (relay replaying stored pings out of order) the last-inserted
  row could be an older location, and same-millisecond ties returned duplicate
  rows per device. timestamp is the sender's fix time, unique per device-ping,
  so the newest actual fix always wins with no ties.
- HeartbeatReceiver retention prune: widen to days.toLong() before multiplying
  so days * 24 * 3600 can't overflow Int for large peer-supplied values.
- MessagingViewModel.onCleared: restore the dropped super.onCleared() call.

+++ app/src/main/java/com/locapeer/subscriber/HeartbeatReceiver.kt
@@ -430,7 +430,9 @@ class HeartbeatReceiver @Inject constructor(
             if (retention.isNotEmpty()) {
                 val now = System.currentTimeMillis()
                 retention.forEach { (deviceId, days) ->
-                    heartbeatDao.deleteOlderThanForDevice(deviceId, now - days * 24 * 3600 * 1000L)
+                    // days.toLong() first: days * 24 * 3600 would overflow Int for large,
+                    // peer-supplied retention values before the trailing 1000L widened it.
+                    heartbeatDao.deleteOlderThanForDevice(deviceId, now - days.toLong() * 24 * 3600 * 1000)
                 }
             }
         } catch (e: Exception) {
